requirements.txt
anyio==3.6.2
Babel==2.12.1
click==8.1.3
docxcompose==1.4.0
docxtpl==0.16.6
fastapi==0.95.1
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
pydantic==1.10.7
python-docx==0.8.11
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.11
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
